split IriCnverterInterface into two different interfaces
